feat: add nearby endpoint & refactor calculateDistance
