+++ backend/src/api/airports/airports.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Get, Param, Query } from '@nestjs/common';
 import { AirportsService } from './airports.service';
 import { airportMaxCountDto } from './dto/airport.max_count.dto';
 import { AirportParams } from './dto/airport.params.dto';
-import { CoordinatesDto } from './dto/coordinates.dto';
+import { CoordinatesDto } from '../../utils/coordinates.dto';
 
 @Controller('api/airports')
 export class AirportsController {
@@ -24,6 +24,15 @@ export class AirportsController {
     ): Promise<AirportCodeDistanceEntity[]> {
       console.log(coordinates);
       return this.airportsService.getDistance(coordinates);
+  }
+
+  @Get('nearby/:max_count?')
+  async getNearby(
+    @Param('max_count') { max_count }: airportMaxCountDto,
+    @Query() coordinates: CoordinatesDto
+    ): Promise<AirportEntity[]> {
+      console.log(coordinates);
+      return this.airportsService.getNearby(coordinates, max_count);
   }
 
   @Get(':iata_code')
